+++ common/src/main/java/com/netflix/conductor/common/constraints/ValidNameConstraint.java
@@ -17,7 +17,11 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import org.springframework.beans.factory.annotation.Value;
+
 import jakarta.validation.Constraint;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 import jakarta.validation.Payload;
 
 import static java.lang.annotation.ElementType.FIELD;
@@ -26,13 +30,11 @@ import static java.lang.annotation.ElementType.FIELD;
  * This constraint class validates following things.
  *
  * <ul>
- *   <li>1. WorkflowDef is valid or not
- *   <li>2. Make sure taskReferenceName used across different tasks are unique
- *   <li>3. Verify inputParameters points to correct tasks or not
+ *   <li>1. Name is valid or not
  * </ul>
  */
 @Documented
-@Constraint(validatedBy = NameValidator.class)
+@Constraint(validatedBy = ValidNameConstraint.NameValidator.class)
 @Target({FIELD})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface ValidNameConstraint {
@@ -42,4 +44,29 @@ public @interface ValidNameConstraint {
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
+
+    class NameValidator implements ConstraintValidator<ValidNameConstraint, String> {
+
+        private static final String NAME_PATTERN = "^[A-Za-z0-9_<>{}#\\s-]+$";
+        public static final String INVALID_NAME_MESSAGE =
+                "Allowed characters are alphanumeric, underscores, spaces, hyphens, and special characters like <, >, {, }, #";
+
+        @Value("${conductor.app.workflow.name-validation.enabled}")
+        private boolean nameValidationEnabled;
+
+        @Override
+        public void initialize(ValidNameConstraint constraintAnnotation) {}
+
+        @Override
+        public boolean isValid(String name, ConstraintValidatorContext context) {
+            boolean valid = name == null || !nameValidationEnabled || name.matches(NAME_PATTERN);
+            if (!valid) {
+                context.disableDefaultConstraintViolation();
+                context.buildConstraintViolationWithTemplate(
+                                "Invalid name '" + name + "'. " + INVALID_NAME_MESSAGE)
+                        .addConstraintViolation();
+            }
+            return valid;
+        }
+    }
 }
